implement: Add Makefile to FPGA implementation environment
Added a Makefile to the FPGA implementation environment.

This environment is used to generate both the .xsa (Xilinx Support
Archive) for the Versal Processing System and the bitstream for the
Programmable Logic (PL). The .xsa archive contains hardware
configuration data required by various tools, such as FSBL,
bootloaders, and software development environments.

The Makefile automates the build process, and currently includes all
necessary steps for synthesizing the Processing System (PS).

Signed-off-by: Daisuke Sasaki <daisuke@spacecubics.com>