implement: Add PL0_REF_CLK as platform clock
Added PL0_REF_CLK as a platform clock.

This clock is required for generating the .xsa file via Export
Hardware Platform. To ensure PL0_REF_CLK is recognized as a platform
clock, it must be connected to proc_sys_reset module.

Since the programmable logic (PL) design has not yet been implemented,
the associated logic is currently a placeholder.

Signed-off-by: Daisuke Sasaki <daisuke@spacecubics.com>